+++ backend/src/index.ts
@@ -6,6 +6,7 @@ import foodsRouter from './routes/foods';
 import logRouter from './routes/log';
 import mealsRouter from './routes/meals';
 import weightRouter from './routes/weight';
+import usersRouter from './routes/users';
 
 dotenv.config();
 
@@ -23,6 +24,7 @@ app.use('/foods', foodsRouter);
 app.use('/log', logRouter);
 app.use('/meals', mealsRouter);
 app.use('/weight', weightRouter);
+app.use('/users', usersRouter);
 
 // Global error handler
 app.use((err: Error, _req: Request, res: Response, _next: NextFunction) => {